Fix issue with leaking connections when downloading remote neurons
Previously, R would run out of connections if more than 124 remote neurons
were requested from a neuronlistfh object (assuming the user had not
previously opened any connections) as the connections were not removed
properly. This is a hack around the issue, rather than a fix for the
underlying cause, but it does mean that subsets of neuronlistfh objects that
require the download of more than 124 remote neurons can now be created.